fead: limitar a los admins a realizar su trabajo y a los usuarios a lo suyo, tanto en el backend como en el frontend

+++ Frontend/pages/(test)/test.vue
@@ -1,5 +1,7 @@
 <template>
-  <div class="min-h-screen bg-gradient-to-br from-gray-50 to-blue-50 flex flex-col">
+  <div
+    class="min-h-screen bg-gradient-to-br from-gray-50 to-blue-50 flex flex-col"
+  >
     <!-- Contenido principal -->
     <div
       :class="[
@@ -9,7 +11,9 @@
     >
       <div class="relative w-full max-w-6xl mx-auto px-6 py-12">
         <!-- Título -->
-        <h1 class="text-4xl md:text-5xl font-bold text-gray-800 mb-12 tracking-tight text-center">
+        <h1
+          class="text-4xl md:text-5xl font-bold text-gray-800 mb-12 tracking-tight text-center"
+        >
           Tests Disponibles
         </h1>
 
@@ -21,13 +25,18 @@
         </div>
 
         <!-- Mensaje cuando no hay tests -->
-        <div v-if="!error && tests.length === 0" class="flex justify-center items-center h-full">
+        <div
+          v-if="!error && tests.length === 0"
+          class="flex justify-center items-center h-full"
+        >
           <p class="text-gray-600 text-center">No hay tests disponibles.</p>
         </div>
 
         <!-- Grid de tests -->
         <div v-if="!error && tests.length > 0" class="relative">
-          <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8 mt-12">
+          <div
+            class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8 mt-12"
+          >
             <div
               v-for="test in tests"
               :key="test.id"
@@ -37,19 +46,32 @@
               ]"
             >
               <!-- Indicador de test incompleto -->
-              <div v-if="test.incompleto" class="absolute top-2 left-2 text-red-500">
+              <div
+                v-if="test.incompleto"
+                class="absolute top-2 left-2 text-red-500"
+              >
                 <span class="text-sm font-semibold">Test Incompleto</span>
               </div>
 
               <!-- Botones de editar, borrar y visualizar -->
-              <div class="absolute top-2 right-2 flex space-x-2">
-                <NuxtLink :to="`/editarTest/${test.id}`" class="text-blue-500 hover:text-blue-700">
+              <div class="absolute top-2 right-2 flex space-x-2"
+              v-if="userRole === 'administrador'">
+                <NuxtLink
+                  :to="`/editarTest/${test.id}`"
+                  class="text-blue-500 hover:text-blue-700"
+                >
                   <PencilIcon class="h-5 w-5" />
                 </NuxtLink>
-                <button @click="borrarTest(test.id)" class="text-red-500 hover:text-red-700">
+                <button
+                  @click="borrarTest(test.id)"
+                  class="text-red-500 hover:text-red-700"
+                >
                   <TrashIcon class="h-5 w-5" />
                 </button>
-                <NuxtLink :to="`/visualizarPreguntas/${test.id}`" class="text-green-500 hover:text-green-700">
+                <NuxtLink
+                  :to="`/visualizarPreguntas/${test.id}`"
+                  class="text-green-500 hover:text-green-700"
+                >
                   <EyeIcon class="h-5 w-5" />
                 </NuxtLink>
               </div>
@@ -72,7 +94,10 @@
                   <span v-if="test.etiqueta === 'Salud'" class="text-green-500">
                     <HeartIcon class="h-5 w-5 inline-block mr-1" />
                   </span>
-                  <span v-if="test.etiqueta === 'Conocimiento'" class="text-yellow-500">
+                  <span
+                    v-if="test.etiqueta === 'Conocimiento'"
+                    class="text-yellow-500"
+                  >
                     <LightBulbIcon class="h-5 w-5 inline-block mr-1" />
                   </span>
                   <span v-if="test.etiqueta === 'Idioma'" class="text-blue-500">
@@ -95,6 +120,7 @@
 
     <!-- Botón flotante para crear test -->
     <NuxtLink
+    v-if="userRole === 'administrador'"
       to="/crearTest"
       class="fixed bottom-6 right-6 bg-blue-600 hover:bg-blue-700 text-white font-bold p-4 rounded-full shadow-lg transition-all duration-300 hover:shadow-lg focus:outline-none focus:shadow-outline"
     >
@@ -115,14 +141,21 @@ import {
   GlobeAltIcon,
 } from "@heroicons/vue/outline";
 import { useRoute } from "#app";
+import { jwtDecode } from "jwt-decode";
 
+
+const userRole = ref(null)
 const tests = ref([]);
 const error = ref(null);
 const runtimeConfig = useRuntimeConfig();
 const apiBaseUrl = runtimeConfig.public.BACKEND_URL;
 const route = useRoute();
 const isAuthPage = computed(() => {
-  return route.path === "/login" || route.path === "/crearUsuario" || route.path === "/verificar-2fa";
+  return (
+    route.path === "/login" ||
+    route.path === "/crearUsuario" ||
+    route.path === "/verificar-2fa"
+  );
 });
 const isNavbarOpen = ref(true); // Estado del Navbar (abierto o cerrado)
 
@@ -132,14 +165,21 @@ const fetchTests = async () => {
 
     for (const test of testsData) {
       try {
-        const preguntasConEstado = await $fetch(`${apiBaseUrl}/preguntas-con-respuestas/test/${test.id}`);
+        const preguntasConEstado = await $fetch(
+          `${apiBaseUrl}/preguntas-con-respuestas/test/${test.id}`
+        );
 
         // Verificar si alguna pregunta está incompleta
-        const tienePreguntasIncompletas = preguntasConEstado.some((pregunta) => !pregunta.completa);
+        const tienePreguntasIncompletas = preguntasConEstado.some(
+          (pregunta) => !pregunta.completa
+        );
 
         test.incompleto = tienePreguntasIncompletas;
       } catch (err) {
-        console.error(`Error al obtener preguntas para el test ${test.id}:`, err);
+        console.error(
+          `Error al obtener preguntas para el test ${test.id}:`,
+          err
+        );
         test.incompleto = true;
       }
     }
@@ -158,6 +198,11 @@ const retryConnection = () => {
 };
 
 const borrarTest = async (id) => {
+  if (userRole.value !== 'administrador') {
+    alert('No tienes permisos para esta acción');
+    return;
+  }
+  
   try {
     await $fetch(`${apiBaseUrl}/test/${id}`, {
       method: "DELETE",
@@ -165,14 +210,27 @@ const borrarTest = async (id) => {
         Authorization: `Bearer ${localStorage.getItem("accessToken")}`,
       },
     });
-    fetchTests(); // Volver a cargar los tests después de borrar
+    fetchTests();
   } catch (err) {
     console.error("Error al borrar el test:", err);
   }
 };
 
+const obtenerRolUsuario = () => {
+  const token = localStorage.getItem("accessToken");
+  if (token) {
+    try {
+      const decoded = jwtDecode(token);
+      userRole.value = decoded.rol.toLowerCase(); // Normaliza a minúsculas
+    } catch (error) {
+      console.error("Error decodificando token:", error);
+      userRole.value = null;
+    }
+  }
+};
 onMounted(() => {
   fetchTests();
+  obtenerRolUsuario();
 });
 useHead({
   title: "Lista de test - TestMaster",
